refactor: move runtime coverage contracts to output crate
